requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-WTF==1.2.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.9.0
Werkzeug==2.3.6
WTForms==3.1.2
zipp==3.16.2
